# docs/swagger.yaml
definitions:
...
  models.VehicleImageReceived:
    properties:
      ref_vehicle_img_side_code:
        example: 1
        type: integer
      vehicle_img_file:
        example: http://vms.pea.co.th/side_image.jpg
        type: string
    type: object
...
  models.VmsTrnReceivedVehicle:
    properties:
      comment_on_received_vehicle:
        example: Minor scratch on bumper
        type: string
      fuel_start:
        example: 50
        type: integer
      mile_start:
        example: 10000
        type: integer
      pickup_datetime:
        example: "2025-03-26T14:30:00Z"
        type: string
      trn_request_uid:
        example: 8bd09808-61fa-42fd-8a03-bf961b5678cd
        type: string
      vehicle_images:
        items:
          $ref: '#/definitions/models.VehicleImageReceived'
        type: array
    type: object
...
info:
  contact: {}
  title: VMS_PLUS
  version: "1.0"
paths:
...
  /api/received-vehicle-user/received-vehicle:
    put:
      consumes:
      - application/json
      description: This endpoint allows to update vehicle pickup.
      parameters:
      - description: VmsTrnReceivedVehicle data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.VmsTrnReceivedVehicle'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Update vehicle pickup for a booking request
      tags:
      - Received-vehicle-user
  /api/received-vehicle-user/request/{trn_request_uid}:
    get:
      consumes:
      - application/json
      description: This endpoint fetches details of a specific booking request using
        its unique identifier (TrnRequestUID).
      parameters:
      - description: TrnRequestUID (trn_request_uid)
        in: path
        name: trn_request_uid
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Retrieve a specific booking request
      tags:
      - Received-vehicle-user
  /api/received-vehicle-user/search-requests:
    get:
      consumes:
      - application/json
      description: Search for requests using a keyword and get the summary of counts
        grouped by request status code
      parameters:
      - description: Search keyword (matches request_no, vehicle_license_plate, vehicle_user_emp_name,
          or work_place)
        in: query
        name: search
        type: string
      - description: Filter by multiple request status codes (comma-separated, e.g.,
          'A,B,C')
        in: query
        name: ref_request_status_code
        type: string
      - description: Filter by start datetime (YYYY-MM-DD format)
        in: query
        name: startdate
        type: string
      - description: Filter by end datetime (YYYY-MM-DD format)
        in: query
        name: enddate
        type: string
      - description: Order by request_no, start_datetime, ref_request_status_code
        in: query
        name: order_by
        type: string
      - description: 'Order direction: asc or desc'
        in: query
        name: order_dir
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of records per page (default: 10)'
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Search booking requests and get summary counts by request status code
      tags:
      - Received-vehicle-user
  /api/received-vehicle-user/travel-card/{trn_request_uid}:
    get:
      consumes:
      - application/json
      description: This endpoint fetches a travel-card of pecific booking request
        using its unique identifier (TrnRequestUID).
      parameters:
      - description: TrnRequestUID (trn_request_uid)
        in: path
        name: trn_request_uid
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      - AuthorizationAuth: []
      summary: Retrieve a travel-card of pecific booking request
      tags:
      - Received-vehicle-user
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-ApiKey
    type: apiKey
  AuthorizationAuth:
    description: Bearer [your Authorization]
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"